moving example to a single file, and adding a counter
