+++ examples/templates/message_passing.aml
