+++ examples/templates/messages.aml
@@ -2,5 +2,4 @@ for value in messages
     text value
 
 padding [top: 1]
-    text "Received messages"
-
+    text "Received messages: " message_count
